requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
openai==1.35.0
anthropic==0.28.0
groq==0.9.0
pydantic==2.7.4
